:recycle: Refactor directory structure
