+++ Dockerfile
@@ -13,4 +13,4 @@ RUN poetry export --with dev > requirements.txt
 RUN pip install -r requirements.txt
 
 EXPOSE 8000
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
