+++ app/api/routes/chat.py
@@ -2,9 +2,15 @@ import os
 import uuid
 
 import chromadb
-from fastapi import FastAPI
-from pydantic import BaseModel, Field
+from fastapi import APIRouter
 from openai import OpenAI
+
+from app.models import (
+    ChatRequest,
+    ChatResponse,
+    LearnRequest,
+)
+
 
 openai_client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
 chromadb_client = chromadb.HttpClient(
@@ -16,29 +22,25 @@ chromadb_client = chromadb.HttpClient(
     tenant=chromadb.config.DEFAULT_TENANT,
     database=chromadb.config.DEFAULT_DATABASE,
 )
-collection = chromadb_client.get_or_create_collection(
+learn_collection = chromadb_client.get_or_create_collection(
     name="learn",
 )
 
-app = FastAPI()
+
+router = APIRouter()
 
 
-@app.get("/")
+@router.get("/")
 def read_root():
     return {"Hello": "World"}
 
 
-class ChatRequest(BaseModel):
-    message: str = Field(description="The message to be sent to the chatbot")
-
-
-class ChatResponse(BaseModel):
-    message: str = Field(description="The message to be responded by the chatbot")
-
-
-@app.post("/query", response_model=ChatResponse, status_code=200)
+@router.post("/query", response_model=ChatResponse, status_code=200)
 def query(chat_request: ChatRequest):
-    rag_result = collection.query(query_texts=[chat_request.message], n_results=1,)[
+    rag_result = learn_collection.query(
+        query_texts=[chat_request.message],
+        n_results=1,
+    )[
         "documents"
     ][0][0]
 
@@ -53,13 +55,9 @@ def query(chat_request: ChatRequest):
     return ChatResponse(message=response.choices[0].message.content)
 
 
-class LearnRequest(BaseModel):
-    text: str = Field(description="The text to be learned by the chatbot")
-
-
-@app.post("/learn", status_code=204)
+@router.post("/learn", status_code=204)
 def learn(learn_request: LearnRequest):
-    collection.add(
+    learn_collection.add(
         documents=[learn_request.text],
         ids=[uuid.uuid4().hex],
     )
