+++ app/main.py
@@ -1,0 +1,7 @@
+from fastapi import FastAPI
+
+from app.api.main import api_router
+
+app = FastAPI()
+
+app.include_router(api_router)
