+++ app/tests/api/routes/test_chat.py
@@ -1,12 +1,7 @@
 from fastapi.testclient import TestClient
 
-from main import app
 
-
-client = TestClient(app)
-
-
-def test_home():
+def test_home(client: TestClient):
     response = client.get("/")
     assert response.status_code == 200
     assert response.json() == {"Hello": "World"}
